+++ modules/canvaskit/canvaskit/types/index.d.ts
@@ -814,7 +814,7 @@ export interface Particles extends EmbindObject<Particles> {
      * Returns the nth uniform from the effect.
      * @param index
      */
-    getUniform(index: number): ParticlesUniform;
+    getUniform(index: number): SkSLUniform;
 
     /**
      * Returns the number of uniforms on the effect.
@@ -860,7 +860,7 @@ export interface Particles extends EmbindObject<Particles> {
     update(now: number): void;
 }
 
-export interface ParticlesUniform {
+export interface SkSLUniform {
     columns: number;
     rows: number;
     /** The index into the uniforms array that this uniform begins. */
@@ -2379,6 +2379,30 @@ export interface RuntimeEffect extends EmbindObject<RuntimeEffect> {
      */
     makeShaderWithChildren(uniforms: Float32Array | number[], isOpaque?: boolean,
                            children?: Shader[], localMatrix?: InputMatrix): Shader;
+
+    /**
+     * Returns the nth uniform from the effect.
+     * @param index
+     */
+    getUniform(index: number): SkSLUniform;
+
+    /**
+     * Returns the number of uniforms on the effect.
+     */
+    getUniformCount(): number;
+
+    /**
+     * Returns the total number of floats across all uniforms on the effect. This is the length
+     * of the uniforms array expected by makeShader. For example, an effect with a single float3
+     * uniform, would return 1 from `getUniformCount()`, but 3 from `getUniformFloatCount()`.
+     */
+    getUniformFloatCount(): number;
+
+    /**
+     * Returns the name of the nth effect uniform.
+     * @param index
+     */
+    getUniformName(index: number): string;
 }
 
 /**